Made changes based on review
