+++ package.json
@@ -28,6 +28,7 @@
     "eslint-config-airbnb-base": "^14.2.1",
     "eslint-plugin-import": "^2.27.5",
     "express": "^4.18.2",
+    "file-loader": "^6.2.0",
     "hint": "^7.1.3",
     "html-webpack-plugin": "^5.5.0",
     "style-loader": "^3.3.1",
