+++ webpack.config.js
@@ -27,8 +27,15 @@ module.exports = {
       },
       {
         test: /\.(png|svg|jpg|jpeg|gif)$/i,
-        type: 'asset/resource',
+        use: [
+          {
+            loader: 'file-loader',
+            options: {
+               limit: 8192,               
+             },
+          },
+      ],
       },
-    ],
+    ],    
   },
 };
